# docs/swagger.yaml
basePath: /v1
definitions:
  dto.InitLoginRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.LoginRequest:
    properties:
      auth_verifier:
        items:
          type: integer
        minItems: 32
        type: array
      email:
        type: string
    required:
    - auth_verifier
    - email
    type: object
  dto.RegisterRequest:
    properties:
      auth_verifier:
        items:
          type: integer
        minItems: 32
        type: array
      email:
        type: string
      enc_private_key:
        items:
          type: integer
        type: array
      public_key:
        items:
          type: integer
        minItems: 32
        type: array
      salt:
        items:
          type: integer
        minItems: 32
        type: array
    required:
    - auth_verifier
    - email
    - enc_private_key
    - public_key
    - salt
    type: object
host: localhost:8080
info:
  contact: {}
  description: Swagger definitions for axosec http api
  title: Axosec Auth API
  version: 0.1.0
paths:
  /auth/login:
    post:
      description: logs user in
      parameters:
      - description: Login request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      responses:
        "200":
          description: OK
      summary: Login
  /auth/login/init:
    post:
      description: initializes login by obtaining users salt
      parameters:
      - description: Login request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.InitLoginRequest'
      responses:
        "200":
          description: OK
      summary: Init Login
  /auth/register:
    post:
      description: registers new user
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      responses:
        "201":
          description: Created
      summary: Register user
  /health:
    get:
      description: returns ok if api up
      responses:
        "200":
          description: OK
      summary: Helthcheck
swagger: "2.0"